Extract FieldMappingEngine from field-matcher
Workstream D, Phase 1 of the architecture-deepening plan.

field-matcher.ts mixed the field-mapping intelligence (the 240-line alias
table and five sibling structures, scoring, overflow routing, the matchFields
algorithm) with a convenience wrapper and the IPC handlers, and pack-builder
reached into it via six dynamic imports.

- Move the matching intelligence verbatim into a new field-mapping-engine.ts,
  a deep module with a small interface (matchFields/getOverflowConfig + the
  field types) that imports nothing from pack-builder.
- field-matcher.ts becomes a thin facade that re-exports the engine surface
  (so every existing './field-matcher' import keeps working) and hosts
  matchSampleToSchema and the fields:* IPC handlers.

Pure move + re-export; behaviour identical, validated by field-matcher.test.ts
(68/68). No pack-builder edits in this phase.
